finalizado pagina de painel de disciplinas com os componentes
