+++ pages/admin/painel/index.js
@@ -1,111 +1,16 @@
 import { AppLayout } from "../../../components/appLayout/appLayout.js"
 import { buttom } from "../../../components/button/buttom.js"
+import { Empty } from "../../../components/empty/empty.js"
 import { Header } from "../../../components/header/header.js"
 import { MainLayout } from "../../../components/mainLayout/mainLayout.js"
 import { Table } from "../../../components/table/table.js"
-import { deleteDisciplinaById, getAllDisciplinasIfProfessorName } from "../../../scripts/fetchDbFunctions.js"
+import { getAllDisciplinasIfProfessorName } from "../../../scripts/fetchDbFunctions.js"
 import { NavBarAdmin } from "../navBarAdm.js"
+import {subjectRegisterBtn, subjectsHeader,tableDataSubjects} from "./subjectsLogic.js"
 
 
-const subjectsHeader = {
-    title: "Disciplinas",
-    subtitle: `${await counterNumberOfSubjectRegisters()} Disciplinas Cadastradas`
-}
 
-const tableDataSubjects = {
-    columns: [
-        {
-            text: "Nome",
-        },
-        {
-            text: "Professor",
-        },
-        {
-            text: "Quiz",
-        },
-        {
-            text: "Ações",
-        }
-    ],
-    rows: await makeArrayRowsWithSubjectsData()
-}
-
-const subjectRegisterBtn = {
-    type: "primary-md",
-    img: "/imgs/createQuizz.svg",
-    text: "Cadastrar",
-    onclick: ()=>{window.location.href = "/pages/admin/register/subjectRegister.html"}
-}
-
-async function makeArrayRowsWithSubjectsData(){
-        
-    const allSubjects = await getAllDisciplinasIfProfessorName()
-
-    if(!allSubjects.disciplinas){
-        return console.log('Lista vazia')
-    }
-
-    let rows = []
-    allSubjects.disciplinas.forEach(disciplina => {
-        let quizesName = []
-        disciplina.quizes.forEach(quiz => {
-            quizesName.push(quiz.nome)
-        })
-        const row = {
-            content1: disciplina.nome,
-            content2: disciplina.prof_id ? disciplina.prof_id.nome : "Não possui professor",
-            action: {
-                content: disciplina.quizes.length,
-                as: "p",
-                hoverItens: quizesName.length > 0 ? quizesName : null,
-            },
-            dialogData: {
-                title: "Tem certeza?",
-                paragraph: `Tem certeza que deseja excluir "${disciplina.nome}"? O processo não poderá ser revertido.`,
-                dialogButtons: [
-                    {
-                        text: "Cancelar",
-                        type: "outline-sm",
-                        onclick(){
-                            const tbody = document.querySelector('.tbody')
-                            const dialog = tbody.querySelector('.dialog')
-                            dialog.remove()
-                            dialog.close()
-                        }
-                    },
-                    {
-                        text: "Eliminar",
-                        type: "destructive-sm",
-                        onclick: async () => {
-                                await deleteDisciplinaById(disciplina._id)
-                                const element = document.getElementById(`${disciplina._id}`) 
-                                element.classList.add('elemento-excluido')
-                                setTimeout(()=>element.remove(), 500)
-
-                        }
-                    },
-                ]
-            },
-            editAnchor: "/pages/admin/dashboard/edit",
-            id: disciplina._id
-        }
-        rows.push(row)
-        
-    });
-
-    return rows
-
-}
-
-async function counterNumberOfSubjectRegisters(){
-    const allSubjects = await getAllDisciplinasIfProfessorName()
-    const length = allSubjects.disciplinas.length
-
-    return length
-}
-
-
-function page(){
+async function subjectsPanelPage(){
     const div = AppLayout()
 
     div.append(NavBarAdmin)
@@ -123,7 +28,19 @@ function page(){
 
     main.append(Table(tableDataSubjects))
 
+    const allSubjects = await getAllDisciplinasIfProfessorName()
+    const emptyDiv = document.createElement('div')
+    emptyDiv.style.height = "100dvh"
+    emptyDiv.style.display = "none"
+    if(!allSubjects.disciplinas){
+        emptyDiv.classList.add('animate-in-login')
+        emptyDiv.style.display = "flex"
+    }
+    emptyDiv.classList.add('empty-subject-div')
+    emptyDiv.append(Empty({title: "Nenhuma disciplina cadastrada"}))
+    main.append(emptyDiv)
+
     document.body.append(div)
 }
 
-page()
+subjectsPanelPage()
