+++ pages/admin/painel/painel.html
@@ -20,7 +20,6 @@
 
 <body>
     <script src="index.js" type="module"></script>
-    <script src="logic.js" type="module"></script>
 </body>
